# backend/openapi.yaml
openapi: 3.1.0
info:
  title: THUIE API
  version: 1.0.0-draft
  description: >-
    REST + WebSocket contract for the THUIE campus alumni info-exchange
    platform. Design skeleton - mirrors BACKEND.md and the PostgreSQL schema in
    PROJECT.md 6. Server implementation deferred to the build phase.
  license: { name: Proprietary }
servers:
  - url: https://api.thuie.example.com/api/v1
    description: Production (placeholder)
  - url: http://localhost:3000/api/v1
    description: Local development

# OpenAPI has no native WebSocket binding, so the realtime surface (BACKEND.md §9)
# is documented here as a vendor extension. Socket.IO server, namespace
# `/realtime`, authenticated with the same access JWT supplied via handshake
# `auth.token`, an `Authorization: Bearer` header, or a `?token=` query.
...
security:
  - bearerAuth: []
paths:
...
  /alumni:
    get: { tags: [Alumni], operationId: listAlumni, summary: Browse alumni, parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: department, schema: { type: string } }, { in: query, name: skill, schema: { type: string } }, { in: query, name: q, schema: { type: string } }, { in: query, name: graduation_year, schema: { type: string } }, { in: query, name: company, schema: { type: string } }, { in: query, name: program, schema: { type: string } }, { in: query, name: country, schema: { type: string } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /alumni/{id}:
    get: { tags: [Alumni], operationId: getAlumni, summary: Alumni profile (contact gated), parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniProfile' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
  /alumni/{id}/connections:
    get: { tags: [Alumni], operationId: listAlumniConnections, summary: Connections of a profile, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AlumniProfile' } } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/alumni-profile:
    get: { tags: [Alumni], operationId: getMyAlumniProfile, summary: Own profile incl drafts/rejected, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniProfile' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Alumni], operationId: createMyAlumniProfile, summary: Create own profile, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AlumniUpdate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniProfile' } } } }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Alumni], operationId: updateMyAlumniProfile, summary: Edit own profile, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AlumniUpdate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniProfile' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /me/alumni-profile/skills:
    put: { tags: [Alumni], operationId: setMySkills, summary: Replace skill set, requestBody: { content: { application/json: { schema: { type: object, properties: { skills: { type: array, items: { type: string } } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/AlumniProfile' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Faculty ----
...
  /admin/review:
    get: { tags: [Admin], operationId: adminReviewQueue, summary: Pending review queue, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: type, schema: { $ref: '#/components/schemas/ModerationTargetType' } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/review/{type}/{id}/approve:
    post: { tags: [Admin], operationId: adminApprove, summary: Approve content, parameters: [ { $ref: '#/components/parameters/ModerationTarget' }, { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: approved }, default: { $ref: '#/components/responses/Error' } } }
  /admin/review/{type}/{id}/reject:
    post: { tags: [Admin], operationId: adminReject, summary: Reject content (reason -> notification), parameters: [ { $ref: '#/components/parameters/ModerationTarget' }, { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { reason: { type: string } } } } } }, responses: { '200': { description: rejected }, default: { $ref: '#/components/responses/Error' } } }
  /admin/takedown:
    post: { tags: [Admin], operationId: adminTakedown, summary: Take content down, requestBody: { content: { application/json: { schema: { type: object, properties: { target_type: { $ref: '#/components/schemas/ModerationTargetType' }, target_id: { type: string }, reason: { type: string } } } } } }, responses: { '200': { description: taken down }, default: { $ref: '#/components/responses/Error' } } }
  /admin/reports:
    get: { tags: [Admin], operationId: adminReports, summary: Reports queue, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: status, schema: { $ref: '#/components/schemas/ReportStatus' } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ReportPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/reports/{id}/resolve:
    post: { tags: [Admin], operationId: adminResolveReport, summary: Resolve report, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { outcome: { type: string, enum: [ignored, deleted, restricted] }, note: { type: string } } } } } }, responses: { '200': { description: resolved }, default: { $ref: '#/components/responses/Error' } } }
  /admin/moderation-history/{type}/{id}:
    get: { tags: [Admin], operationId: adminModerationHistory, summary: moderation_actions timeline, parameters: [ { $ref: '#/components/parameters/ModerationTarget' }, { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ModerationAction' } } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/keywords:
    get: { tags: [Admin], operationId: adminKeywords, summary: List keywords, responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Keyword' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Admin], operationId: adminCreateKeyword, summary: Create keyword (busts cache), requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/KeywordCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Keyword' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/keywords/{id}:
    patch: { tags: [Admin], operationId: adminUpdateKeyword, summary: Update keyword, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/KeywordCreate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Keyword' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Admin], operationId: adminDeleteKeyword, summary: Delete keyword, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: deleted }, default: { $ref: '#/components/responses/Error' } } }
  /admin/users:
    get: { tags: [Admin], operationId: adminUsers, summary: Search users, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: q, schema: { type: string } }, { in: query, name: role, schema: { $ref: '#/components/schemas/Role' } }, { in: query, name: status, schema: { $ref: '#/components/schemas/UserStatus' } }, { in: query, name: program, schema: { type: string } }, { in: query, name: nationality, schema: { type: string } }, { in: query, name: sort, schema: { type: string, enum: [name, program, department, role, status, created_at] } }, { in: query, name: order, schema: { type: string, enum: [asc, desc] } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/UserPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Admin], operationId: adminCreateUser, summary: Provision a user (admin-supplied temporary password, starts active), requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/users/{id}:
    patch: { tags: [Admin], operationId: adminUpdateUser, summary: Restrict / ban / restore (status only), parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, required: [status], properties: { status: { $ref: '#/components/schemas/UserStatus' } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    put: { tags: [Admin], operationId: adminPutUser, summary: Edit user profile details, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/UserAdminUpdate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Admin], operationId: adminDeleteUser, summary: Delete a user, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: deleted }, default: { $ref: '#/components/responses/Error' } } }
  /admin/role-strategies:
    get: { tags: [Admin], operationId: adminRoleStrategies, summary: List role strategies, responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/RoleStrategy' } } } } }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Admin], operationId: adminUpdateRoleStrategy, summary: Update flags (admin_super only), requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/RoleStrategy' } } } }, responses: { '200': { description: ok }, default: { $ref: '#/components/responses/Error' } } }
...
  /admin/feedback:
    get: { tags: [Admin], operationId: adminFeedbackQueue, summary: Feedback triage queue, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: status, schema: { type: string, enum: [open, answered, closed] } } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Feedback' } } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/feedback/{id}:
    patch: { tags: [Admin], operationId: adminReplyFeedback, summary: Reply to / transition a feedback thread, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { reply: { type: string, maxLength: 2000 }, status: { type: string, enum: [open, answered, closed] } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Feedback' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/operation-logs:
    get: { tags: [Admin], operationId: adminOperationLogs, summary: Admin audit log, parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/OperationLogPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Admin], operationId: adminClearOperationLogs, summary: Clear the audit log, responses: { '204': { description: cleared }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Auth: reset / verify ----
...
  /me/alumni-profile/visibility-request:
    post: { tags: [Alumni], operationId: requestVisibilityChange, summary: Ask admin to change visibility, requestBody: { content: { application/json: { schema: { type: object, properties: { visibility: { $ref: '#/components/schemas/Visibility' }, note: { type: string } } } } } }, responses: { '202': { description: submitted for admin review }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Reports (generic) ----
...
  /admin/users/{id}/convert:
    post: { tags: [Admin], operationId: adminConvertUser, summary: student -> graduate conversion, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { graduation_year: { type: string }, force: { type: boolean } } } } } }, responses: { '200': { description: converted, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /admin/users/convert:
    post: { tags: [Admin], operationId: adminConvertUsersBatch, summary: Batch conversion, requestBody: { content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { type: object, properties: { user_id: { type: string }, graduation_year: { type: string } } } } } } } } }, responses: { '200': { description: done }, default: { $ref: '#/components/responses/Error' } } }
  /admin/stats:
    get: { tags: [Admin], operationId: adminStats, summary: Dashboard stats, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Stats' } } } }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Ops ----
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    ModerationTarget: { in: path, name: type, required: true, schema: { $ref: '#/components/schemas/ModerationTargetType' } }
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    Id: { in: path, name: id, required: true, schema: { type: string } }
...
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    Cursor: { in: query, name: cursor, schema: { type: string }, description: Opaque keyset cursor }
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found (or hidden by visibility)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Role: { type: string, enum: [student, graduate, admin, admin_super] }
    UserStatus: { type: string, enum: [unverified, active, posting_restricted, banned, deleted] }
...
    ProfileStatus: { type: string, enum: [draft, pending, approved, rejected] }
    Visibility: { type: string, enum: [student_only, all, admin_only] }
...
    ProfileSource: { type: string, enum: [school, user] }
...
    KeywordAction: { type: string, enum: [block, manual_review] }
...
    ReportStatus: { type: string, enum: [open, resolved_ignored, resolved_deleted, resolved_restricted] }
...
    ModerationTargetType: { type: string, enum: [info_post, forum_post, comment, alumni_profile] }
    ReportTargetType: { type: string, enum: [info_post, forum_post, comment, alumni_profile, user] }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine code }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
        errors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
    CursorMeta: { type: object, properties: { nextCursor: { type: [string, 'null'] }, limit: { type: integer } } }
    OffsetMeta: { type: object, properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer } } }
...
    Department: { type: object, properties: { id: { type: string }, code: { type: string }, name_zh: { type: string }, name_en: { type: string }, faculty: { type: [string, 'null'] } } }
    User: { type: object, properties: { id: { type: string }, name: { type: string }, role: { $ref: '#/components/schemas/Role' }, status: { $ref: '#/components/schemas/UserStatus' }, student_id: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, email: { type: [string, 'null'] }, department: { $ref: '#/components/schemas/Department' }, grade_year: { type: [string, 'null'] }, graduation_year: { type: [string, 'null'] }, avatar_url: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, wechat: { type: [string, 'null'], description: GET /me only }, whatsapp: { type: [string, 'null'], description: GET /me only }, linkedin: { type: [string, 'null'], description: GET /me only }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' }, is_online: { type: boolean }, created_at: { type: string, format: date-time } } }
...
    UserCreate: { type: object, required: [name, role, password], properties: { name: { type: string }, role: { $ref: '#/components/schemas/Role' }, password: { type: string, description: admin-supplied temporary password }, student_id: { type: string }, email: { type: string }, phone: { type: string }, department_id: { type: string }, program: { type: string }, nationality: { type: string }, grade_year: { type: string }, graduation_year: { type: string } } }
    UserAdminUpdate: { type: object, properties: { name: { type: string }, email: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, student_id: { type: [string, 'null'] }, department_id: { type: [string, 'null'] }, program: { type: [string, 'null'] }, nationality: { type: [string, 'null'] }, grade_year: { type: [string, 'null'] }, graduation_year: { type: [string, 'null'] }, role: { $ref: '#/components/schemas/Role' } } }
    RoleStrategy: { type: object, properties: { role: { $ref: '#/components/schemas/Role' }, can_view_info: { type: boolean }, can_view_internal: { type: boolean }, can_view_alumni: { type: boolean }, can_view_forum: { type: boolean }, can_submit_info: { type: boolean }, can_post_forum: { type: boolean }, can_comment: { type: boolean }, can_create_profile: { type: boolean } } }
...
    AlumniProfile: { type: object, properties: { id: { type: string }, display_name: { type: string }, department: { $ref: '#/components/schemas/Department' }, graduation_year: { type: [string, 'null'] }, program: { type: [string, 'null'] }, industry: { type: [string, 'null'] }, country: { type: [string, 'null'] }, city: { type: [string, 'null'] }, work_title: { type: [string, 'null'] }, company: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, skills: { type: array, items: { type: string } }, source: { $ref: '#/components/schemas/ProfileSource' }, status: { $ref: '#/components/schemas/ProfileStatus' }, visibility: { $ref: '#/components/schemas/Visibility' }, email: { type: [string, 'null'], description: omitted unless the owner account's email_visibility + role allowed }, phone: { type: [string, 'null'], description: omitted unless the owner account's phone_visibility + role allowed }, wechat: { type: [string, 'null'], description: omitted unless the owner account's wechat_visibility + role allowed }, whatsapp: { type: [string, 'null'], description: omitted unless the owner account's whatsapp_visibility + role allowed }, linkedin: { type: [string, 'null'], description: omitted unless the owner account's linkedin_visibility + role allowed }, completion_percent: { type: integer }, version: { type: integer }, avatar_url: { type: [string, 'null'] } } }
    AlumniUpdate: { type: object, properties: { display_name: { type: string }, department_id: { type: string }, graduation_year: { type: string }, program: { type: string }, industry: { type: string }, country: { type: string }, city: { type: string }, work_title: { type: string }, company: { type: string }, bio: { type: string }, visibility: { $ref: '#/components/schemas/Visibility' } } }
...
    Feedback: { type: object, properties: { id: { type: string }, content: { type: string }, reply: { type: [string, 'null'] }, status: { type: string, enum: [open, answered, closed] }, created_at: { type: string, format: date-time } } }
    Keyword: { type: object, properties: { id: { type: string }, word: { type: string }, action: { $ref: '#/components/schemas/KeywordAction' }, enabled: { type: boolean } } }
    KeywordCreate: { type: object, required: [word, action], properties: { word: { type: string, maxLength: 100 }, action: { $ref: '#/components/schemas/KeywordAction' }, enabled: { type: boolean, default: true } } }
    Report: { type: object, properties: { id: { type: string }, reporter: { $ref: '#/components/schemas/User' }, target_type: { $ref: '#/components/schemas/ReportTargetType' }, target_id: { type: string }, reason: { type: string }, status: { $ref: '#/components/schemas/ReportStatus' }, result_note: { type: [string, 'null'] }, created_at: { type: string, format: date-time } } }
...
    ModerationAction: { type: object, properties: { id: { type: string }, target_type: { $ref: '#/components/schemas/ModerationTargetType' }, target_id: { type: string }, actor: { oneOf: [ { $ref: '#/components/schemas/User' }, { type: 'null' } ], description: null = keyword engine / system }, action: { type: string, enum: [submitted, approved, rejected, taken_down, removed, resubmitted] }, reason: { type: [string, 'null'] }, created_at: { type: string, format: date-time } } }
    OperationLog: { type: object, properties: { id: { type: string }, admin: { $ref: '#/components/schemas/User' }, action: { type: string }, target_type: { type: [string, 'null'] }, target_id: { type: [string, 'null'] }, detail: { type: object, additionalProperties: true }, ip: { type: [string, 'null'] }, created_at: { type: string, format: date-time } } }
    ReviewQueueItem: { type: object, properties: { target_type: { $ref: '#/components/schemas/ModerationTargetType' }, target_id: { type: string }, title: { type: string }, submitted_at: { type: string, format: date-time }, author: { $ref: '#/components/schemas/User' } } }
...
    Stats: { type: object, properties: { total_users: { type: integer }, total_infos: { type: integer }, total_posts: { type: integer }, total_profiles: { type: integer }, pending_review: { type: integer }, open_reports: { type: integer }, oldest_open_report_hours: { type: integer }, dau: { type: integer }, wau: { type: integer }, signups_7d: { type: integer } } }
...
    AlumniPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/AlumniProfile' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...
    ReviewPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/ReviewQueueItem' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
    ReportPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Report' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
    UserPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/User' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }
    OperationLogPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/OperationLog' } }, meta: { $ref: '#/components/schemas/OffsetMeta' } } }